backend/requirements.txt
fastapi
uvicorn
motor
beanie
python-dotenv
google-generativeai
numpy
opencv-python-headless
pydantic
httpx
